+++ R/get_mask.R
@@ -32,49 +32,5 @@
 #'}
 #'
 get_mask <- function(bf, format = "SpatRaster") {
-
-  format <- tolower(format)
-  format <- switch(format,
-                   "spatrast" = "spatraster",
-                   "terra" = "spatraster",
-                   "matrix" = "numeric",
-                   "array" = "numeric",
-                   "raster.data.frame" = "dataframe",
-                   "data.frame" = "dataframe",
-                   format)
-
-  stopifnot("Format must be one of 'SpatRaster', 'numeric', or 'dataframe'" =
-              format %in% c("spatraster", "numeric", "dataframe"))
-
-  if (format == "spatraster") {
-    m <- bf$geom$mask
-    r <- terra::rast(m, extent = bf$geom$ext, crs = bf$geom$crs)
-    names(r) <- "mask"
-    return(r)
-  }
-
-  if (format == "numeric") {
-    return(bf$geom$mask)
-  }
-
-  if (format == "dataframe") {
-    # Data frame for use with ggplot2 geom_raster() it will
-    # have one row per cell in the full raster
-
-    # Get x, y, i, and mask columns for all cells in the full raster
-    #  i will be NA for cells that aren't in the mask
-    rows <- seq_len(nrow(bf))
-    cols <- seq_len(ncol(bf))
-    df <- expand.grid(rows, cols)
-    names(df) <- c("row", "col")
-    df$x <- col_to_x(df$col, bf)
-    df$y <- row_to_y(df$row, bf)
-    df$i <- rc_to_i(df$row, df$col, bf)
-    mv <- match(1:n_active(bf), df$i)
-    df$mask <- FALSE
-    df$mask[mv] <- TRUE
-    return(df)
-  }
-
-  stop("Unrecognized format.") # should not ever get here
+  format_raster_data(bf$geom$mask, bf, format, "mask")
 }